Moving tests around to make it more manageable.
